To match analysis, see desc
It doesn't change anything for adding, but this way the user can only edit nickname and watering period. plant name, latin name, image and the ideal environment can't be changed acording to the use case description.  Imo it would make sense to be able to edit image as well as the plant grows.